refactor: 테스트용 지하철역과 노선을 재사용 가능하도록 변경
